% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.character.outcome}
\alias{as.character.outcome}
\title{Convert outcome vector to character vector}
\usage{
\method{as.character}{outcome}(x, ...)
}
\arguments{
\item{x}{Outcome vector.}

\item{...}{Ignored, kept for S3 consistency.}
}
\value{
A character vector.
}
\description{
Convert outcome vector to character vector
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

