% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.outcome}
\alias{as.data.frame.outcome}
\title{Convert outcome vector to data frame}
\usage{
\method{as.data.frame}{outcome}(x, ...)
}
\arguments{
\item{x}{Outcome vector.}

\item{...}{Ignored, kept for S3 consistency.}
}
\value{
The data frame underlying the outcome vector.
}
\description{
Convert outcome vector to data frame
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

