% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.matrix.outcome}
\alias{as.matrix.outcome}
\title{Convert outcome vector to matrix}
\usage{
\method{as.matrix}{outcome}(x, ...)
}
\arguments{
\item{x}{Outcome vector.}

\item{...}{Ignored, kept for S3 consistency.}
}
\value{
The data frame underlying the outcome vector converted to a matrix.
  Event types are converted to integers.
}
\description{
Convert outcome vector to matrix
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

