% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_procedure.r
\name{as.modeling_procedure}
\alias{as.modeling_procedure}
\title{Coerce to modeling procedure}
\usage{
as.modeling_procedure(x, ...)
}
\arguments{
\item{x}{modeling procedure.}

\item{...}{Ignored (kept for S3 consistency).}
}
\value{
Modeling procedure
}
\description{
Coerce to modeling procedure
}
\examples{
as.modeling_procedure("lda")
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
