% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.outcome}
\alias{as.outcome}
\title{Convert object to outcome vector}
\usage{
as.outcome(x, ...)
}
\arguments{
\item{x}{Generic object.}

\item{...}{Sent to class methods.}
}
\value{
A vector of class \code{\link{outcome}}.
}
\description{
Converts objects to class \code{\link{outcome}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

