% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.outcome.Surv}
\alias{as.outcome.Surv}
\title{Convert Surv vector to outcome vector}
\usage{
\method{as.outcome}{Surv}(x, ...)
}
\arguments{
\item{x}{Surv vector.}

\item{...}{Ignored, kept for S3 consistency.}
}
\value{
A vector of class \code{\link{outcome}}.
}
\description{
Convert Surv vector to outcome vector
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

