% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.Surv.outcome}
\alias{as.Surv.outcome}
\title{Convert outcome vector to Surv vector}
\usage{
\method{as.Surv}{outcome}(x, main = 1, censor = NA, ...)
}
\arguments{
\item{x}{Outcome vector.}

\item{main}{Surv only supports one event type. This argument controls
which type that will be kept, all others are considered censoring.}

\item{censor}{A vector of event types to consider censoring.}

\item{...}{Ignored, kept for S3 consistency.}
}
\value{
A vector of class \code{\link{Surv}}. All event types other than
  the main event are considered censoring.
}
\description{
Convert outcome vector to Surv vector
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

