% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.Surv}
\alias{as.Surv}
\title{Convert object to Surv vector}
\usage{
as.Surv(x, ...)
}
\arguments{
\item{x}{Generic object.}

\item{...}{Sent to class methods.}
}
\value{
A vector of class \code{\link{Surv}}.
}
\description{
Converts objects to class \code{\link{Surv}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

