% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.Surv.Surv}
\alias{as.Surv.Surv}
\title{Trivial function}
\usage{
\method{as.Surv}{Surv}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Ignored.}
}
\value{
A vector of class \code{\link{Surv}}.
}
\description{
Defined in case \code{\link{as.Surv}} is called on a \code{\link{Surv}}
object.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

