% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dim.outcome}
\alias{dim.outcome}
\title{Dimension of an outcome vector}
\usage{
\method{dim}{outcome}(x)
}
\arguments{
\item{x}{Outcome vector.}
}
\value{
A vector with number of observations and number of variables, of
  which there is always two, time and event.
}
\description{
Dimension of an outcome vector
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

