% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.fit.caret}
\alias{emil.fit.caret}
\title{Fit a model using the caret package}
\usage{
emil.fit.caret(x, y, ...)
}
\arguments{
\item{x}{Descriptors.}

\item{y}{Response.}

\item{...}{Sent to \code{\link[caret]{train}}.}
}
\description{
Fit a model using the caret package
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

