% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.fit.lda}
\alias{emil.fit.lda}
\title{Fit linear discriminant}
\usage{
emil.fit.lda(x, y, ...)
}
\arguments{
\item{x}{Dataset, numerical matrix with observations as rows.}

\item{y}{Class labels, factor.}

\item{...}{Sent to \code{\link{lda}}.}
}
\value{
Fitted linear discriminant.
}
\description{
Wrapper for the MASS package implementation.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.predict.lda}},
  \code{\link{modeling.procedure}}
}

