% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.fit.lm}
\alias{emil.fit.lm}
\title{Fit a linear model fitted with ordinary least squares}
\usage{
emil.fit.lm(x, y, formula = y ~ ., ...)
}
\arguments{
\item{x}{Descriptors.}

\item{y}{Response, numeric.}

\item{formula}{See \code{\link{lm}}.}

\item{...}{Sent to \code{\link{lm}}.}
}
\value{
Fitted linear model.
}
\description{
Based on \code{\link{lm}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.predict.lm}},
  \code{\link{modeling.procedure}}
}

