% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.fit.qda}
\alias{emil.fit.qda}
\title{Fit quadratic discriminant.}
\usage{
emil.fit.qda(x, y, ...)
}
\arguments{
\item{x}{Dataset, numerical matrix with observations as rows.}

\item{y}{Class labels, factor.}

\item{...}{Sent to \code{\link{qda}}.}
}
\value{
Fitted QDA.
}
\description{
Wrapper for the MASS package implementation.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.predict.qda}},
  \code{\link{modeling.procedure}}
}

