% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.fit.randomForest}
\alias{emil.fit.randomForest}
\title{Fit random forest.}
\usage{
emil.fit.randomForest(x, y, importance = FALSE, ...)
}
\arguments{
\item{x}{Dataset, numerical matrix with observations as rows.}

\item{y}{Class labels, factor.}

\item{importance}{Whether to calculate permuted OOB error as a variable
importance measure, see \code{\link[randomForest]{importance.randomForest}}. Set to
\code{FALSE} to speed up computation.}

\item{...}{Sent to \code{\link[randomForest]{randomForest}}.}
}
\value{
Fitted random forest.
}
\description{
Directly calling the \code{randomForest} package implementation. See
\code{\link[randomForest]{randomForest}} for parameter specification.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.predict.randomForest}},
  \code{\link{emil.vimp.randomForest}}, \code{\link{modeling.procedure}}
}

