% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.predict.cforest}
\alias{emil.predict.cforest}
\title{Predict with conditional inference forest}
\usage{
emil.predict.cforest(object, x, at, ...)
}
\arguments{
\item{object}{Fitted \code{cforest} classifier, as returned by
\code{\link{emil.fit.cforest}}.}

\item{x}{New data to be used for predictions.}

\item{at}{Time point to evaluate survival curves at. If omitted it is set
to the last observed time point.}

\item{...}{Sent to \code{\link[party]{treeresponse}}·}
}
\value{
The predicted chance of survival.
}
\description{
Prediction function for models fitted with \code{\link{emil.fit.cforest}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.fit.cforest}},
  \code{\link{modeling.procedure}}
}

