% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.predict.glmnet}
\alias{emil.predict.glmnet}
\title{Predict using generalized linear model with elastic net regularization}
\usage{
emil.predict.glmnet(object, x, s, ...)
}
\arguments{
\item{object}{Fitted model.}

\item{x}{New data to be predicted.}

\item{s}{Regularization parameter lambda.}

\item{...}{Sent to \code{\link[glmnet]{predict.glmnet}}.}
}
\value{
A list with elements:
\itemize{
    \item{\code{pred}: Factor of predicted class memberships.}
    \item{\code{prob}: Data frame of predicted class probabilities.}
}
}
\description{
Due to the way \code{\link[glmnet]{glmnet}} is implemented, the regularization alpha
can not be modified after the model is fitted.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.fit.glmnet}},
  \code{\link{modeling.procedure}}
}

