% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.predict.lm}
\alias{emil.predict.lm}
\title{Prediction using linear model}
\usage{
emil.predict.lm(object, x, ...)
}
\arguments{
\item{object}{Fitted classifier produced by \code{\link{emil.fit.lm}}.}

\item{x}{Dataset to be predicted upon.}

\item{...}{Sent to \code{\link{predict.lm}}}
}
\value{
A list with elements:
\itemize{
    \item{\code{pred}: Vector of predicted response.}
}
}
\description{
Prediction using linear model
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.fit.lm}},
  \code{\link{modeling.procedure}}
}

