% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.predict.pamr}
\alias{emil.predict.pamr}
\title{Prediction using nearest shrunken centroids.}
\usage{
emil.predict.pamr(object, x, threshold, thres.fun, ...)
}
\arguments{
\item{object}{Fitted classifier.}

\item{x}{Dataset of observations to be classified.}

\item{threshold}{Threshold to use for classification. This argument is only
needed if you want to override the value set during model fitting.}

\item{thres.fun}{Threshold selection function. Only needed if you want to
override the function set during model fitting.}

\item{...}{Sent to \code{\link[pamr]{pamr.predict}}.}
}
\value{
A list with elements:
\itemize{
    \item{\code{pred}: Factor of predicted class memberships.}
    \item{\code{prob}: Data frame of predicted class probabilities.}
}
}
\description{
In case multiple thresholds give the same error the largest one is chosen
(i.e. the one keeping the fewest features).
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.fit.pamr}},
  \code{\link{emil.vimp.pamr}}, \code{\link{modeling.procedure}}
}

