% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.predict.qda}
\alias{emil.predict.qda}
\title{Prediction using already trained classifier.}
\usage{
emil.predict.qda(object, x, ...)
}
\arguments{
\item{object}{Fitted classifier as produced by \code{\link{batch.model}}.}

\item{x}{Dataset of observations to be classified.}

\item{...}{Sent to \code{\link{predict.qda}}.}
}
\value{
A list with elements:
\itemize{
    \item{\code{pred}: Factor of predicted class memberships.}
    \item{\code{prob}: Data frame of predicted class probabilities.}
}
}
\description{
Wrapper for the MASS package implementation.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.fit.qda}},
  \code{\link{modeling.procedure}}
}

