% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.predict.randomForest}
\alias{emil.predict.randomForest}
\title{Prediction using random forest.}
\usage{
emil.predict.randomForest(object, x, ...)
}
\arguments{
\item{object}{Fitted model.}

\item{x}{Dataset of observations to be classified.}

\item{...}{Ignored}
}
\value{
When used for classification, a list with elements:
\itemize{
    \item{\code{pred}: Factor of predicted class memberships.}
    \item{\code{prob}: Data frame of predicted class probabilities.}
}

When used for regression, a list with the element:
\itemize{
    \item{\code{pred}: Vector of predicted response.}
}
}
\description{
Prediction using random forest.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.fit.randomForest}},
  \code{\link{emil.vimp.randomForest}}, \code{\link{modeling.procedure}}
}

