% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.vimp.pamr}
\alias{emil.vimp.pamr}
\title{Variable importance of nearest shrunken centroids.}
\usage{
emil.vimp.pamr(object, threshold, thres.fun, ...)
}
\arguments{
\item{object}{Fitted pamr classifier}

\item{threshold}{Threshold to use for classification. This argument is only
needed if you want to override the value set during model fitting.}

\item{thres.fun}{Threshold selection function. Only needed if you want to
override the function set during model fitting.}

\item{...}{Sent to \code{\link[pamr]{pamr.predict}}.}
}
\value{
A matrix of variable importance scores where the rows represent
  variables and the columns represent classes.
}
\description{
Calculated as the absolute difference between the overall centroid and a
class-wise shrunken centroid (which is the same for both classes except sign).
}
\details{
In case multiple thresholds give the same error the largest one is chosen
(i.e. the one keeping the fewest features).
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.fit.pamr}},
  \code{\link{emil.predict.pamr}}, \code{\link{modeling.procedure}}
}

