% Generated by roxygen2 (4.0.1): do not edit by hand
\name{emil.vimp.randomForest}
\alias{emil.vimp.randomForest}
\title{Variable importance of random forest.}
\usage{
emil.vimp.randomForest(object, type = 1, ...)
}
\arguments{
\item{object}{Fitted randomForest classifier}

\item{type}{Importance can be assessed in two ways:
\describe{
  \item{1.}{Permuted out-of-bag prediction error (default). This can only be
         used if the classifier was fitted with argument
         \code{importance=TRUE} which is default.}
  \item{2.}{Total decrease in node impurity.}
}}

\item{...}{Ignored.}
}
\value{
An importance vector with elements corresponding to variables.
}
\description{
Variable importance of random forest.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{emil.fit.randomForest}},
  \code{\link{emil.predict.randomForest}}, \code{\link{modeling.procedure}}
}

