% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.r
\name{emil_list_method}
\alias{emil_list_method}
\title{List all available methods}
\usage{
emil_list_method(pos = search())
}
\arguments{
\item{pos}{Location to search in, see \code{\link{ls}}.}
}
\value{
A data frame.
}
\description{
This function searches all attached packages for methods compatible with the
emil framework.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

