% Generated by roxygen2 (4.0.1): do not edit by hand
\name{[.outcome}
\alias{[.outcome}
\title{Extract}
\usage{
\method{[}{outcome}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{Outcome vector.}

\item{i}{Index.}

\item{j}{Column index, if given \code{x} is treated as a data frame.}

\item{drop}{See \code{\link{Extract}}}
}
\value{
A subset of \code{x}.
}
\description{
Extract
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

