% Generated by roxygen2 (4.0.1): do not edit by hand
\name{factor.events}
\alias{factor.events}
\title{Get events on factor form}
\usage{
factor.events(x, time, censor.label = "no event", keep)
}
\arguments{
\item{x}{Outcome vector.}

\item{time}{Time point to evaluate at.}

\item{censor.label}{What to label the absence of an event with.}

\item{keep}{Event types to keep, defaults to all.}
}
\value{
A factor of events.
}
\description{
By default all events are returned regardless of when they occurred. By
setting \code{time} only the events that has occurred up until then will be
returned, and cases with shorter follow-up times and no event will be marked
as censored.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

