% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.r
\name{fit_caret}
\alias{fit_caret}
\title{Fit a model using the \pkg{caret} package}
\usage{
fit_caret(x, y, ...)
}
\arguments{
\item{x}{Descriptors.}

\item{y}{Response.}

\item{...}{Sent to \code{\link[caret]{train}}.}
}
\description{
Fit a model using the \pkg{caret} package
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
