% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{fit_lda}
\alias{fit_lda}
\title{Fit linear discriminant}
\usage{
fit_lda(x, y, ...)
}
\arguments{
\item{x}{Dataset, numerical matrix with observations as rows.}

\item{y}{Class labels, factor.}

\item{...}{Sent to \code{\link{lda}}.}
}
\value{
Fitted linear discriminant.
}
\description{
Wrapper for the \pkg{MASS} package implementation.
}
\seealso{
\code{\link{emil}}, \code{\link{predict_lda}},
  \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
