% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.r
\name{fit_lm}
\alias{fit_lm}
\title{Fit a linear model fitted with ordinary least squares}
\usage{
fit_lm(x, y, formula = y ~ ., ...)
}
\arguments{
\item{x}{Descriptors.}

\item{y}{Response, numeric.}

\item{formula}{See \code{\link{lm}}.}

\item{...}{Sent to \code{\link{lm}}.}
}
\value{
Fitted linear model.
}
\description{
Based on \code{\link{lm}}.
}
\seealso{
\code{\link{emil}}, \code{\link{predict_lm}},
  \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
