% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e1071.r
\name{fit_naive_bayes}
\alias{fit_naive_bayes}
\title{Fit a naive Bayes classifier}
\usage{
fit_naive_bayes(x, y, ...)
}
\arguments{
\item{x}{Dataset, observations as rows.}

\item{y}{Response vector.}

\item{...}{Send to \code{\link[e1071]{naiveBayes}}.}
}
\description{
Fit a naive Bayes classifier
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
