% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomForest.r
\name{fit_randomForest}
\alias{fit_randomForest}
\title{Fit random forest.}
\usage{
fit_randomForest(x, y, prediction = FALSE, ...)
}
\arguments{
\item{x}{Dataset, numerical matrix with observations as rows.}

\item{y}{Class labels, factor.}

\item{prediction}{Whether to calculate permuted OOB error as a feature
prediction measure, see \code{\link[randomForest]{predict.randomForest}}.
Set to \code{FALSE} to speed up computation.}

\item{...}{Sent to \code{\link[randomForest]{randomForest}}.}
}
\value{
Fitted random forest.
}
\description{
Directly calling the \code{randomForest} package implementation. See
\code{\link[randomForest]{randomForest}} for parameter specification.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{predict_randomForest}},
  \code{\link{importance_randomForest}}, \code{\link{modeling_procedure}}
}

