% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e1071.r
\name{fit_svm}
\alias{fit_svm}
\title{Fit a support vector machine}
\usage{
fit_svm(x, y, probability = TRUE, ranges, ...)
}
\arguments{
\item{x}{Dataset, observations as rows.}

\item{y}{Response vector.}

\item{probability}{If \code{FALSE} support for class probability estimation
is not included in the fitted model. This may save some computation time.}

\item{ranges}{Parameter ranges to tune over. Sent to
\code{\link[e1071]{tune.svm}}.}

\item{...}{Sent to \code{\link[e1071]{svm}}.}
}
\description{
Fit a support vector machine
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
