% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape-result.r
\name{get_performance}
\alias{get_performance}
\title{Extract prediction performance}
\usage{
get_performance(result, format = c("wide", "long"))
}
\arguments{
\item{result}{Modeling result, as returned by \code{\link{evaluate}}.}

\item{format}{Table format of the output. See
\url{http://en.wikipedia.org/wiki/Wide_and_narrow_data} for more info.}
}
\value{
Data frame.
}
\description{
Extract prediction performance
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
