% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.r
\name{get_response}
\alias{get_response}
\title{Extract the response from a data set}
\usage{
get_response(x, y)
}
\arguments{
\item{x}{Data set features.}

\item{y}{Response vector or any other type of objects that describe how to
extract the response vector from \code{x}.}
}
\value{
A response vector.
}
\description{
Extract the response from a data set
}
\examples{
identical(iris$Species, get_response(iris, "Species"))
identical(iris$Sepal.Length, get_response(iris, Sepal.Length ~ .))
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
