% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppWrappers.r
\name{has_na}
\alias{has_na}
\title{Check if an object contains missing values}
\usage{
has_na(x)
}
\arguments{
\item{x}{Vector or matrix.}
}
\value{
Logical scalar that is \code{TRUE} if \code{x} contains one or more
  missing values and \code{FALSE} if \code{x} contains no missing values.
}
\description{
Check if an object contains missing values
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

