% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.r
\name{importance_glmnet}
\alias{importance_glmnet}
\alias{importance_ridge_regression}
\alias{importance_lasso}
\title{Feature importance extractor for elastic net models}
\usage{
importance_glmnet(object, s, ...)

importance_ridge_regression(object, s, ...)

importance_lasso(object, s, ...)
}
\arguments{
\item{object}{Fitted elastic net model, as produced by
\code{\link{fit_glmnet}}.}

\item{s}{Regularization parameter lambda.}

\item{...}{Sent to \code{\link[glmnet]{predict.glmnet}}.}
}
\value{
A feature imortance data frame.
}
\description{
Feature importance extractor for elastic net models
}
\seealso{
\code{\link{emil}}, \code{\link{fit_glmnet}},
  \code{\link{predict_glmnet}}, \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
