% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.r
\name{importance_randomForest}
\alias{importance_randomForest}
\title{Feature importance of random forest.}
\usage{
importance_randomForest(object, type = 1, ...)
}
\arguments{
\item{object}{Fitted randomForest classifier}

\item{type}{Importance can be assessed in two ways:
\describe{
  \item{1.}{Permuted out-of-bag prediction error (default). This can only be
         used if the classifier was fitted with argument
         \code{prediction=TRUE} which is default.}
  \item{2.}{Total decrease in node impurity.}
}}

\item{...}{Ignored.}
}
\value{
An prediction vector with elements corresponding to variables.
}
\description{
Feature importance of random forest.
}
\seealso{
\code{\link{emil}}, \code{\link{fit_randomForest}},
  \code{\link{predict_randomForest}}, \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
