% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.r
\name{indent}
\alias{indent}
\title{Increase indentation}
\usage{
indent(base, indent)
}
\arguments{
\item{base}{Base indentation level of the function printing the message.}

\item{indent}{Extra indentation of this message.}
}
\value{
An integer that can be used to specify the indentation level of
  messages printed with \code{\link{log_message}}.
}
\description{
Increase indentation
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
