% Generated by roxygen2 (4.0.1): do not edit by hand
\name{index.fit}
\alias{index.fit}
\alias{index.test}
\title{Convert a fold to row indexes of fittdng or test set}
\usage{
index.fit(fold, allow.oversample = TRUE)

index.test(fold)
}
\arguments{
\item{fold}{A fold of a resampling scheme.}

\item{allow.oversample}{Whether or not to allow individual observation to
exist in multiple copies in the training set. This is typically not the
case, but can be used when a class is underrepresented in the data set.}
}
\value{
An integer vector of row indexes.
}
\description{
Convert a fold to row indexes of fittdng or test set
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{resample}}
}

