% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.r
\name{index_fit}
\alias{index_fit}
\alias{index_test}
\title{Convert a fold to row indexes of fittdng or test set}
\usage{
index_fit(fold, allow_oversample = TRUE)

index_test(fold)
}
\arguments{
\item{fold}{A fold of a resampling scheme.}

\item{allow_oversample}{Whether or not to allow individual observation to
exist in multiple copies in the training set. This is typically not the
case, but can be used when a class is underrepresented in the data set.}
}
\value{
An integer vector of row indexes.
}
\description{
Convert a fold to row indexes of fittdng or test set
}
\seealso{
\code{\link{emil}}, \code{\link{resample}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
