% Generated by roxygen2 (4.0.1): do not edit by hand
\name{integer.events}
\alias{integer.events}
\title{Return events in integer form}
\usage{
integer.events(x, ...)
}
\arguments{
\item{x}{Outcome vector.}

\item{...}{Sent to \code{\link{factor.events}}.}
}
\value{
Integer vector.
}
\description{
Basically calls \code{\link{factor.events}} and converts to integer. No
event is coded with 0 and the other event types with > 0.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

