% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.blank}
\alias{is.blank}
\title{Wrapper for several methods to test if a variable is empty}
\usage{
is.blank(x, false.triggers = FALSE)
}
\arguments{
\item{x}{A variable.}

\item{false.triggers}{Whether \code{FALSE} should be considered as empty.}
}
\value{
Logical telling if variable is blank.
}
\description{
This is mainly an internal function but as other dependent packages also
use it sometimes and it generally is quite handy to have it is exported for
public use.
}
\examples{
is.blank(NULL)
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

