% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.na.outcome}
\alias{is.na.outcome}
\title{Check for missing values}
\usage{
\method{is.na}{outcome}(x)
}
\arguments{
\item{x}{Outcome vector.}
}
\value{
A logical vector with \code{TRUE} where an outcome is missing.
}
\description{
Check for missing values
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

