% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.outcome}
\alias{is.outcome}
\title{Test if object is of class outcome}
\usage{
is.outcome(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} if \code{x} is of class \code{outcome}. \code{FALSE}
  otherwise.
}
\description{
Test if object is of class outcome
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

