% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.writable}
\alias{is.writable}
\title{Test if a directory is writable}
\usage{
is.writable(path = ".")
}
\arguments{
\item{path}{The path to the directory}
}
\value{
Logical indicating if the directory is writable
}
\description{
Test if a directory is writable
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

