% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{is_blank}
\alias{is_blank}
\title{Wrapper for several methods to test if a variable is empty}
\usage{
is_blank(x, false_triggers = FALSE)
}
\arguments{
\item{x}{A variable.}

\item{false_triggers}{Whether \code{FALSE} should be considered as empty.}
}
\value{
Logical telling if variable is blank.
}
\description{
This is mainly an internal function but as other dependent packages also
use it sometimes and it generally is quite handy to have it is exported for
public use.
}
\examples{
is_blank(NULL)
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
