% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppWrappers.r
\name{is_constant}
\alias{is_constant}
\title{Check if an object contains more than one unique value}
\usage{
is_constant(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector or matrix.}

\item{na.rm}{Whether to ignore missing values.}
}
\value{
Logical scalar that is \code{TRUE} if \code{x} contains more than one
  unique value and \code{FALSE} if not.

A logical scalar that is \code{TRUE} if \code{x} contains more than
  one unique value and \code{FALSE} otherwise. In case \code{x} contains
  missing values \code{NA} is returned if \code{na_rm = FALSE}. If there are
  no non-missing values \code{NA} is always returned.
}
\description{
Check if an object contains more than one unique value
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
