% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{is_multi_procedure}
\alias{is_multi_procedure}
\title{Detect if modeling results contains multiple procedures}
\usage{
is_multi_procedure(result)
}
\arguments{
\item{result}{Modeling results, as returned by \code{\link{evaluate}}.}
}
\value{
Logical scalar.
}
\description{
Detect if modeling results contains multiple procedures
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
