% Generated by roxygen2 (4.0.1): do not edit by hand
\name{length.outcome}
\alias{length.outcome}
\title{Length of an outcome vector}
\usage{
\method{length}{outcome}(x)
}
\arguments{
\item{x}{Outcome vector.}
}
\value{
Length.
}
\description{
Length of an outcome vector
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

