% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{list_method}
\alias{list_method}
\title{List all available methods}
\usage{
list_method(pos = search())
}
\arguments{
\item{pos}{Location to search in, see \code{\link{ls}}.}
}
\value{
A data frame.
}
\description{
This function searches all attached packages for methods compatible with the
\pkg{emil} framework.
}
\examples{
list_method()
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
