% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{na_index}
\alias{na_index}
\title{Support function for identifying missing values}
\usage{
na_index(data)
}
\arguments{
\item{data}{Fitting and testing data sets, as returned by
\code{\link{pre_split}}.}
}
\value{
Data frame containing row and column indices of missing values or
  \code{NULL} if the data doesn't contain any.
}
\description{
Support function for identifying missing values
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

