% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{name_procedure}
\alias{name_procedure}
\title{Get names for modeling procedures}
\usage{
name_procedure(procedure)
}
\arguments{
\item{procedure}{List of modeling procedures.}
}
\value{
A character vector of suitable non-duplicate names.
}
\description{
Get names for modeling procedures
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
