% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neg.gmpa}
\alias{neg.gmpa}
\title{Negative geometric mean of class specific predictive accuracy}
\usage{
neg.gmpa(true, pred)
}
\arguments{
\item{true}{See \code{\link{error.fun}}.}

\item{pred}{See \code{\link{error.fun}}.}
}
\value{
A numeric scalar.
}
\description{
When dealing with imbalanced classification problem, i.e. where the class
sizes are very different, small classes tend to be overlooked when tuning
parameters by optimizing error rate. Blagus and Lusa (2013) suggested to
remedy the problem by using this performance measure instead.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\references{
Blagus, R., & Lusa, L. (2013).
\emph{Improved shrunken centroid classifiers for high-dimensional class-imbalanced data.}
BMC bioinformatics, 14, 64.
doi:10.1186/1471-2105-14-64
}
\seealso{
\code{\link{error.fun}}
}

