% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-functions.r
\name{neg_gmpa}
\alias{neg_gmpa}
\title{Negative geometric mean of class specific predictive accuracy}
\usage{
neg_gmpa(truth, prediction, na.rm = FALSE)
}
\arguments{
\item{truth}{See \code{\link{error_fun}}.}

\item{prediction}{See \code{\link{error_fun}}.}

\item{na.rm}{Whether to remove missing values or not.}
}
\value{
A numeric scalar.
}
\description{
When dealing with imbalanced classification problem, i.e. where the class
sizes are very different, small classes tend to be overlooked when tuning
parameters by optimizing error rate. Blagus and Lusa (2013) suggested to
remedy the problem by using this performance measure instead.
}
\references{
Blagus, R., & Lusa, L. (2013).
\emph{Improved shrunken centroid classifiers for high-dimensional class-imbalanced data.}
BMC bioinformatics, 14, 64.
doi:10.1186/1471-2105-14-64
}
\seealso{
\code{\link{error_fun}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
