% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nice.require}
\alias{nice.require}
\title{Load a package and offer to install if missing}
\usage{
nice.require(pkg, reason = "is required")
}
\arguments{
\item{pkg}{Package name.}

\item{reason}{A status message that informs the user why the package is
needed.}
}
\value{
Nothing
}
\description{
If running R in interactive mode, the user is prompted.
}
\examples{
nice.require("base", "is required to do anything at all")
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

