% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{nice_axis}
\alias{nice_axis}
\title{Plots an axis the way an axis should be plotted.}
\usage{
nice_axis(..., las = 1, lwd = 0, lwd.ticks = par("lwd"), lend = 2)
}
\arguments{
\item{...}{Sent to \code{\link{axis}}.}

\item{las}{Rotation of axis labels. Always horizontal by default.}

\item{lwd}{Width of the line drawn along the plot area. Omitted by default
since it overlaps with \code{\link{box}} and causes it to look thicker
where the axis is.}

\item{lwd.ticks}{Width of the tick lines. These are kept by default.}

\item{lend}{Line endings, see \code{\link{par}}.}
}
\description{
Plots an axis the way an axis should be plotted.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
