% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{nice_box}
\alias{nice_box}
\title{Plots a box around a plot}
\usage{
nice_box(lend = 2, ljoin = 1, ...)
}
\arguments{
\item{lend}{Line ending style, see \code{\link{par}}. Defaults to square.}

\item{ljoin}{Line joint style, see \code{\link{par}}. Defaults to mitre,
i.e. 90 degree corners in this case.}

\item{...}{Sent to \code{\link{box}}.}
}
\description{
Plots a box around a plot
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
