% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{nice_require}
\alias{nice_require}
\title{Load a package and offer to install if missing}
\usage{
nice_require(pkg, reason)
}
\arguments{
\item{pkg}{Package name.}

\item{reason}{A status message that informs the user why the package is
needed.}
}
\value{
Nothing
}
\description{
If running R in interactive mode, the user is prompted for installing
missing packages. If running in batch mode an error is thrown.
}
\examples{
nice_require("base", "is required to do anything at all")
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
