% Generated by roxygen2 (4.0.1): do not edit by hand
\name{p.value.coxph}
\alias{p.value.coxph}
\title{Extract p-value from a Cox proportional hazards model}
\usage{
\method{p.value}{coxph}(x, log.p = FALSE, test = c("logrank", "wald",
  "likelihood"), ...)
}
\arguments{
\item{x}{Fitted \code{\link[survival]{coxph}} model.}

\item{log.p}{Whether to return the logarithm of the p-value.}

\item{test}{What test to calculate. \code{"likelihood"} is short for means
likelihood ratio test.}

\item{...}{Ignored. Kept for S3 consistency.}
}
\value{
p-value.
}
\description{
Based on \code{\link{summary.coxph}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{p.value}}
}

