% Generated by roxygen2 (4.0.1): do not edit by hand
\name{p.value.cuminc}
\alias{p.value.cuminc}
\title{Extract p-value from a cumulative incidence estimation}
\usage{
\method{p.value}{cuminc}(x, log.p = FALSE, ...)
}
\arguments{
\item{x}{Fitted \code{\link[cmprsk]{cuminc}} estimate.}

\item{log.p}{Whether to return the logarithm of the p-value.}

\item{...}{Ignored. Kept for S3 consistency.}
}
\value{
p-value.
}
\description{
This is also known as Gray's test.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{p.value}}
}

