% Generated by roxygen2 (4.0.1): do not edit by hand
\name{p.value}
\alias{p.value}
\title{Extraction of p-value from a statistical test}
\usage{
p.value(x, log.p = FALSE, ...)
}
\arguments{
\item{x}{Test, i.e. a fitted object of a supported type.}

\item{log.p}{Whether to return the logarithm of the p-value.}

\item{...}{Sent to class method.}
}
\value{
p-value.
}
\description{
These calculations are written in such a way that they avoid rounding off errors
that plague the \code{survival} and \code{cmprsk} packages.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{p.value.crr}}, \code{\link{p.value.survdiff}},
  \code{\link{p.value.cuminc}}
}

