% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-curve.r
\name{plot.learning_curve}
\alias{plot.learning_curve}
\title{Plot results from learning curve analysis}
\usage{
\method{plot}{learning_curve}(x, ..., summaries = list(mean = mean,
  `95-percentile` = function(x) quantile(x, 0.95)))
}
\arguments{
\item{x}{Results from \code{\link{learning_curve}}.}

\item{...}{Ignored, kept for S3 consistency.}

\item{summaries}{Named list of summary functions that can reduce a vector of
performance estimates to a single quantity.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot results from learning curve analysis
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
