% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.outcome}
\alias{plot.outcome}
\title{Plot outcome vector}
\usage{
\method{plot}{outcome}(x, y, segments = TRUE, flip = FALSE,
  legendpos = "topright", ...)
}
\arguments{
\item{x}{outcome vector.}

\item{y}{Y-values.}

\item{segments}{Whether to draw horizontal segments.}

\item{flip}{Flip the plot to show time on y.}

\item{legendpos}{Position of legend, see \code{\link{legend}}. Set to NA or
NULL to suppress legend.}

\item{...}{Sent to \code{\link{plot}}.}
}
\description{
Plot outcome vector
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

