% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival.r
\name{plot.Surv}
\alias{plot.Surv}
\title{Plot Surv vector}
\usage{
\method{plot}{Surv}(x, y, segments = TRUE, flip = FALSE,
  legendpos = "topright", ...)
}
\arguments{
\item{x}{\code{\link{Surv}} vector.}

\item{y}{Y-values.}

\item{segments}{Whether to draw horizontal segments.}

\item{flip}{Flip the plot to show time on y.}

\item{legendpos}{Position of legend, see \code{\link{legend}}. Set to NA or
NULL to suppress legend.}

\item{...}{Sent to \code{\link{plot}}.}
}
\description{
Plot Surv vector
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

