% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pre.pamr}
\alias{pre.pamr}
\title{PAMR adapted dataset pre-processing}
\usage{
pre.pamr(x, y, fold, pre.process = pre.split, ...)
}
\arguments{
\item{x}{Dataset.}

\item{y}{Response vector.}

\item{fold}{A logical vector with \code{FALSE} for fitting observations,
\code{TRUE} for test observations and \code{NA} for observations not
to be included.}

\item{pre.process}{A pre-processing function to be wrapped.}

\item{...}{Sent to \code{pre.process}.}
}
\value{
A list with fitting and testing sets, formatted the way pamr wants
  them.
}
\description{
The predict framework is designed to work with dataset where rows correspond
to observations and columns to descriptors. PAMR wants it the other way, and
also to have the fitting set response vector supplied in a list with the
descriptors. This function applies a standard pre-processing function and
then reformats the result to satisfy PAMR.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{pre.process}}
}

