% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.r
\name{pre_impute}
\alias{pre_impute}
\alias{pre_impute_median}
\alias{pre_impute_mean}
\title{Basic imputation}
\usage{
pre_impute(data, fun, ...)

pre_impute_median(data)

pre_impute_mean(data)
}
\arguments{
\item{data}{Fitting and test datasets, as returned by \code{\link{pre_split}}
or any other standard pre-processing function.}

\item{fun}{Function for calculating imputation values. Should take a vector
and return a scalar.}

\item{...}{Sent to \code{fun}.}
}
\value{
A pair of fitting and testing datasets.
}
\description{
This solution is optimized for the scenario that the dataset is very large
but only contains missing values in a small number of columns.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
