% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{pre_log_message}
\alias{pre_log_message}
\title{Print log message during pre-processing}
\usage{
pre_log_message(data, ...)
}
\arguments{
\item{data}{Pre-processed data set.}

\item{...}{Sent to \code{\link{log_message}}}
}
\value{
The same data set as inputted. This only purpose of this function is
  to print a log message as a side effect.
}
\description{
Print log message during pre-processing
}
\seealso{
\code{\link{pre_process}}, \code{\link{log_message}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
