% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.r
\name{predict.model}
\alias{predict.model}
\title{Predict the response of unknown observations}
\usage{
\method{predict}{model}(object, x, ..., .verbose = FALSE)
}
\arguments{
\item{object}{Fitted model.}

\item{x}{Data set with observations whose response is to be predicted.}

\item{...}{Sent to the procedure's prediction function.}

\item{.verbose}{Whether to print an activity log.}
}
\value{
See the documentation of procedure's method.
}
\description{
Predict the response of unknown observations
}
\examples{
mod <- fit("lda", x=iris[-5], y=iris$Species)
prediction <- predict(mod, iris[-5])
}
\seealso{
\code{\link{emil}}, \code{\link{modeling_procedure}},
  \code{\link{evaluate}},\code{\link{fit}}, \code{\link{tune}},
  \code{\link{get_importance}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
