% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.modeling.procedure}
\alias{predict.modeling.procedure}
\title{Predict the response of unknown observations}
\usage{
\method{predict}{modeling.procedure}(object, model, x, ...)
}
\arguments{
\item{object}{Modeling procedure.}

\item{model}{Fitted model.}

\item{x}{Data set with observations whose response is to be predicted.}

\item{...}{Sent to the procedure's prediction function.}
}
\value{
See the documentation of procedure's method.
}
\description{
Predict the response of unknown observations
}
\examples{
proc <- modeling.procedure("lda")
mod <- fit(proc, x=iris[-5], y=iris$Species)
pred <- predict(proc, mod, iris[-5])
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{modeling.procedure}},
  \code{\link{evaluate.modeling}},\code{\link{fit}}, \code{\link{tune}},
  \code{\link{vimp}}
}

