% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cforest.r
\name{predict_cforest}
\alias{predict_cforest}
\title{Predict with conditional inference forest}
\usage{
predict_cforest(object, x, at, ...)
}
\arguments{
\item{object}{Fitted \code{cforest} classifier, as returned by
\code{\link{fit_cforest}}.}

\item{x}{New data to be used for predictions.}

\item{at}{Time point to evaluate survival curves at. If omitted it is set
to the last observed time point.}

\item{...}{Sent to \code{\link[party]{treeresponse}}·}
}
\value{
The predicted chance of survival.
}
\description{
Prediction function for models fitted with \code{\link{fit_cforest}}.
}
\seealso{
\code{\link{emil}}, \code{\link{fit_cforest}},
  \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
