% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.r
\name{predict_coxph}
\alias{predict_coxph}
\title{Predict using Cox proportional hazards model}
\usage{
predict_coxph(object, x, ...)
}
\arguments{
\item{object}{Fitted model, as returned by \code{\link{fit_coxph}}.}

\item{x}{Observations whose response is to be predicted.}

\item{...}{Sent to \code{\link{predict.coxph}}.}
}
\description{
Predict using Cox proportional hazards model
}
\seealso{
\code{\link{fit_coxph}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
