% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.r
\name{predict_glmnet}
\alias{predict_glmnet}
\title{Predict using generalized linear model with elastic net regularization}
\usage{
predict_glmnet(object, x, s, ...)
}
\arguments{
\item{object}{Fitted model.}

\item{x}{New data to be predicted.}

\item{s}{Regularization parameter lambda.}

\item{...}{Sent to \code{\link[glmnet]{predict.glmnet}}.}
}
\value{
A list with elements:
\itemize{
    \item{\code{prediction}: Factor of predicted class memberships.}
    \item{\code{probability}: Data frame of predicted class probabilities.}
}
}
\description{
Due to the way \code{\link[glmnet]{glmnet}} is implemented, the regularization alpha
can not be modified after the model is fitted.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{fit_glmnet}},
  \code{\link{modeling_procedure}}
}

