% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.r
\name{predict_lm}
\alias{predict_lm}
\title{Prediction using linear model}
\usage{
predict_lm(object, x, ...)
}
\arguments{
\item{object}{Fitted classifier produced by \code{\link{fit_lm}}.}

\item{x}{Dataset to be predicted upon.}

\item{...}{Sent to \code{\link{predict.lm}}}
}
\value{
A list with elements:
\itemize{
    \item{\code{prediction}: Vector of predicted response.}
}
}
\description{
Prediction using linear model
}
\seealso{
\code{\link{emil}}, \code{\link{fit_lm}},
  \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
