% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamr.r
\name{predict_pamr}
\alias{predict_pamr}
\title{Prediction using nearest shrunken centroids.}
\usage{
predict_pamr(object, x, threshold, thres_fun, ...)
}
\arguments{
\item{object}{Fitted classifier.}

\item{x}{Dataset of observations to be classified.}

\item{threshold}{Threshold to use for classification. This argument is only
needed if you want to override the value set during model fitting.}

\item{thres_fun}{Threshold selection function. Only needed if you want to
override the function set during model fitting.}

\item{...}{Sent to \code{\link[pamr]{pamr.predict}}.}
}
\value{
A list with elements:
\itemize{
    \item{\code{prediction}: Factor of predicted class memberships.}
    \item{\code{probability}: Data frame of predicted class probabilities.}
}
}
\description{
In case multiple thresholds give the same error the largest one is chosen
(i.e. the one keeping the fewest features).
}
\seealso{
\code{\link{emil}}, \code{\link{fit_pamr}},
  \code{\link{importance_pamr}}, \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
