% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.r
\name{predict_randomForest}
\alias{predict_randomForest}
\title{Prediction using random forest.}
\usage{
predict_randomForest(object, x, ...)
}
\arguments{
\item{object}{Fitted model.}

\item{x}{Dataset of observations to be classified.}

\item{...}{Ignored}
}
\value{
When used for classification, a list with elements:
\itemize{
    \item{\code{prediction}: Factor of predicted class memberships.}
    \item{\code{probability}: Data frame of predicted class probabilities.}
}

When used for regression, a list with the element:
\itemize{
    \item{\code{prediction}: Vector of predicted response.}
}
}
\description{
Prediction using random forest.
}
\seealso{
\code{\link{emil}}, \code{\link{fit_randomForest}},
  \code{\link{importance_randomForest}}, \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
