% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.r
\name{predict_rpart}
\alias{predict_rpart}
\title{Predict using a fitted decision tree}
\usage{
predict_rpart(object, x)
}
\arguments{
\item{object}{Fitted decision tree.}

\item{x}{New data whose response is to be predicted.}
}
\value{
Predictions. The exact form depends on the type of application
  (classification or regression)
}
\description{
Predict using a fitted decision tree
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
