% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e1071.r
\name{predict_svm}
\alias{predict_svm}
\title{Predict using support vector machine}
\usage{
predict_svm(object, x, probability = object$compprob, statistic = TRUE, ...)
}
\arguments{
\item{object}{Fitted SVM, as produced by \code{\link{fit_svm}}.}

\item{x}{Data set to predict response for.}

\item{probability}{Whether to calculate class probabilities.}

\item{statistic}{Whether to return the raw classification statistics.}

\item{...}{Sent to \code{\link[e1071:predict.svm]{predict}}.}
}
\description{
Predict using support vector machine
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
