% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.outcome}
\alias{print.outcome}
\title{Print outcome vector}
\usage{
\method{print}{outcome}(x, quote = FALSE, ...)
}
\arguments{
\item{x}{Outcome vector.}

\item{quote}{Logical, whether to print quotation marks.}

\item{...}{Ignored, kept for S3 consistency.}
}
\value{
Nothing, only prints the vector to stdout.
}
\description{
Print outcome vector
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

