% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{print.preprocessed_data}
\alias{print.preprocessed_data}
\title{Print method for pre-processed data}
\usage{
\method{print}{preprocessed_data}(x, ...)
}
\arguments{
\item{x}{Pre-processed data, as produced by \code{\link{pre_split}}.}

\item{...}{Ignored, kept for S3 consistency.}
}
\value{
Nothing
}
\description{
Print method for pre-processed data
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
