% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.r
\name{pvalue.cuminc}
\alias{pvalue.cuminc}
\title{Extract p-value from a cumulative incidence estimation}
\usage{
\method{pvalue}{cuminc}(x, log_p = FALSE, ...)
}
\arguments{
\item{x}{Fitted \code{\link[cmprsk]{cuminc}} estimate.}

\item{log_p}{Whether to return the logarithm of the p-value.}

\item{...}{Ignored. Kept for S3 consistency.}
}
\value{
p-value.
}
\description{
This is also known as Gray's test.
}
\seealso{
\code{\link{pvalue}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
