% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.r
\name{pvalue}
\alias{pvalue}
\title{Extraction of p-value from a statistical test}
\usage{
pvalue(x, log_p = FALSE, ...)
}
\arguments{
\item{x}{Test, i.e. a fitted object of a supported type.}

\item{log_p}{Whether to return the logarithm of the p-value.}

\item{...}{Sent to class method.}
}
\value{
p-value.
}
\description{
These calculations are written in such a way that they avoid rounding off errors
that plague the \pkg{survival} and \pkg{cmprsk} packages.
}
\seealso{
\code{\link{pvalue.crr}}, \code{\link{pvalue.survdiff}},
  \code{\link{pvalue.cuminc}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
