% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-functions.r
\name{trivial_error_rate}
\alias{trivial_error_rate}
\title{Calculate the trivial error rate}
\usage{
trivial_error_rate(truth)
}
\arguments{
\item{truth}{True class labels.}
}
\description{
Simply predicting the most common class for all test set observations can be
a decievingly successful strategy in terms of error rate. This function shows
what error rate such a strategy would result in.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
