% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{validate_data}
\alias{validate_data}
\title{Validate a pre-processed data set}
\usage{
validate_data(data)
}
\arguments{
\item{data}{Pre-processed data set.}
}
\value{
Nothing, only throws an error or prints a completion message.
}
\description{
While writing and debugging pre-processing functions this function can be
useful to confirm that the resulting data sets fulfills the necessary
requirements.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
