% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{vlines}
\alias{vlines}
\alias{hlines}
\title{Add vertical or horizontal lines to a plot}
\usage{
vlines(x, lend = 1, ...)

hlines(y, lend = 1, ...)
}
\arguments{
\item{x}{Coordinates of vertical lines.}

\item{lend}{Line ending style, see \code{\link{par}}.}

\item{...}{Sent to \code{\link{segments}}.}

\item{y}{Coordinates of horizontal lines.}
}
\description{
Add vertical or horizontal lines to a plot
}
\examples{
plot(0:10, 0:10, type="n")
hlines(0:4*2.5, col="#dddddd")
points(0:10, 0:10)
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
