% Generated by roxygen2 (4.0.1): do not edit by hand
\name{warn.once}
\alias{reset.warn.once}
\alias{warn.once}
\title{Print a warning message if not printed earlier}
\usage{
warn.once(id, ...)

reset.warn.once()
}
\arguments{
\item{id}{Warning message id. This is used internally to refer to the
message.}

\item{...}{Sent to \code{\link{warning}}.}
}
\description{
To avoid flooding the user with identical warning messages, this function
keeps track of which have already been shown.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

