% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weighted.error.rate}
\alias{weighted.error.rate}
\title{Weighted error rate}
\usage{
weighted.error.rate(x)
}
\arguments{
\item{x}{Cost matrix or factor response vector.}
}
\value{
An error function.
}
\description{
If different types of errors are associated with different costs a weighted
error function might be more appropriate than the standard.
}
\details{
This function is not in itself an error function, but used to generate error
functions. Either supply a predefined cost matrix or a response
vector for a classification problem to define it automatically.

The automatically generated cost matrix will generate an error of 0 if all
predictions are correct, 1 if all predictions are incorrect and 0.5 if all
predictions are the same (regardless of class, i.e. if one class is smaller
it will be given a higher misclassification cost).
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

