% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{fill}
\alias{fill}
\alias{na_fill}
\title{Replace values with something else}
\usage{
fill(x, pattern, replacement, invert = FALSE)

na_fill(x, replacement)
}
\arguments{
\item{x}{Variable containing NAs.}

\item{pattern}{The values in \code{x} to be replaced. Can also be a
function.}

\item{replacement}{The value which is to replace the values matching
\code{pattern}.}

\item{invert}{Whether to fill all values except the ones matching
\code{pattern}.}
}
\value{
An imputed version of \code{x}.
}
\description{
Replace values with something else
}
\examples{
fill(1:10, function(x) x \%\% 2 == 1, 0)
na_fill(c(1,2,NA,4,5), 3)
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
