% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.r
\name{fit_caret}
\alias{fit_caret}
\title{Fit a model using the \pkg{caret} package}
\usage{
fit_caret(x, y, ...)
}
\arguments{
\item{x}{Descriptors.}

\item{y}{Response.}

\item{...}{Sent to \code{\link[caret]{train}}.}
}
\description{
Fit a model using the \pkg{caret} package
}
\references{
Kuhn M (2008). “Building Predictive Models in R Using the caret Package.
  Journal of Statistical Software, 28(5), 1–26. doi:10.18637/jss.v028.i05.
  caret: Classification and Regression Training. R package version 6.0-79,
  URL https://CRAN.R-project.org/package=caret.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
