% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.r
\name{fit_randomForest}
\alias{fit_randomForest}
\title{Fit random forest.}
\usage{
fit_randomForest(x, y, ...)
}
\arguments{
\item{x}{Dataset, numerical matrix with observations as rows.}

\item{y}{Class labels, factor.}

\item{...}{Sent to \code{\link[randomForest]{randomForest}}.}
}
\value{
Fitted random forest.
}
\description{
Directly calling the \pkg{randomForest} package implementation. See
\code{\link[randomForest]{randomForest}} for parameter specification.
}
\references{
Breiman L (2001). Random Forests. Machine Learning, 45(1), 5–32.
  doi:10.1023/a:1010933404324.
}
\seealso{
\code{\link{emil}}, \code{\link{predict_randomForest}},
  \code{\link{importance_randomForest}}, \code{\link{modeling_procedure}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
