% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.r
\name{fit_rpart}
\alias{fit_rpart}
\title{Fit a decision tree}
\usage{
fit_rpart(x, y, ...)
}
\arguments{
\item{x}{Data set (features).}

\item{y}{Response.}

\item{...}{Sent to \code{\link{rpart}}.}
}
\value{
A fitted decision tree.
}
\description{
Fit a decision tree
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
