% Generated by roxygen2 (4.0.1): do not edit by hand
\name{impute}
\alias{impute}
\alias{impute.knn}
\alias{impute.median}
\title{Regular imputation}
\usage{
impute.knn(x, k = 0.05, distmat = "auto")

impute.median(x)
}
\arguments{
\item{x}{Dataset.}

\item{k}{Number of nearest neighbors to use.}

\item{distmat}{Distance matrix.}
}
\value{
An imputed matrix.
}
\description{
If you want to impute, build model and predict you should use
\code{\link{pre.impute.median}} or \code{\link{pre.impute.knn}}.
This function imputes using all observations
without caring about cross-validation folds.
}
\details{
For additional information on the parameters see \code{\link{pre.impute.knn}}
and \code{\link{pre.impute.median}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{emil}}, \code{\link{pre.process}},
  \code{\link{pre.impute.knn}}, \code{\link{pre.impute.median}}
}

