% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamr.r
\name{pre_pamr}
\alias{pre_pamr}
\title{PAMR adapted dataset pre-processing}
\usage{
pre_pamr(data)
}
\arguments{
\item{data}{Fitting and testing data sets, as returned by
\code{\link{pre_split}}.}
}
\value{
A list with fitting and testing sets, formatted the way pamr wants
  them.
}
\description{
The predict framework is designed to work with dataset where rows correspond
to observations and columns to descriptors. PAMR wants it the other way, and
also to have the fitting set response vector supplied in a list with the
descriptors. This function applies a standard pre-processing function and
then reformats the result to satisfy PAMR.
}
\details{
\code{pre_pamr} must be run last if chained with other pre-processing
functions, since it substantially reshapes the data.
}
\seealso{
\code{\link{emil}}, \code{\link{pre_process}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
