% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caret.r
\name{predict_caret}
\alias{predict_caret}
\title{Predict using a caret method}
\usage{
predict_caret(...)
}
\arguments{
\item{...}{Sent to \code{\link{predict}} that forwards it to the
appropriate predict function in the caret package.}
}
\description{
This is not guaranteed to work with all caret methods. If it doesn't work
for a particular method, the user will need to rewrite it.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

