% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/e1071.r
\name{predict_naive_bayes}
\alias{predict_naive_bayes}
\title{Predict using naive Bayes model}
\usage{
predict_naive_bayes(object, x)
}
\arguments{
\item{object}{Fitted naive Bayes model.}

\item{x}{Data set to predict response for}
}
\description{
Predict using naive Bayes model
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link[e1071:predict.svm]{predict}}.
}

