% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.r
\name{subresample}
\alias{subresample}
\title{Generate resampling subschemes}
\usage{
subresample(fold, y)
}
\arguments{
\item{fold}{A resampling scheme or fold to use to define the sub scheme(s).}

\item{y}{The observations used to create the resampling scheme. See
\code{\link{resample}} for details.}
}
\value{
A resampling scheme.
}
\description{
A subscheme is a resampling scheme that only includes observations in the
training set of a fold. This function
automatically fetches the type and parameters of the prototype fold and use
them to generate the subscheme.
}
\examples{
cv <- resample("holdout", y=1:12, test_fraction=1/4, nfold=3)
inner.cv <- subresample(cv, y=1:12)
}
\seealso{
\code{\link{emil}}, \code{\link{resample}}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
